Try another approach to implement clean architecture
