+++ lib/domain-on-working/auth/repositories/auth_repository.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:new_chatify/core/error/failures.dart';
-import 'package:new_chatify/features/auth/domain/entities/user_entitiy.dart';
+import 'package:new_chatify/domain-on-working/shared/entities/user_entitiy.dart';
 
 abstract class AuthRepository {
   Future<Either<Failure, UserEntity>> signUp(String email, String password);
