+++ lib/domain-on-working/auth/usecases/signin_usecase.dart
@@ -1,17 +1,17 @@
 import 'package:dartz/dartz.dart';
 import 'package:new_chatify/core/error/failures.dart';
 import 'package:new_chatify/core/usecases/usecase.dart';
-import 'package:new_chatify/features/auth/domain/entities/user_entitiy.dart';
-import 'package:new_chatify/features/auth/domain/repositories/auth_repository.dart';
+import 'package:new_chatify/domain-on-working/shared/entities/user_entitiy.dart';
+import 'package:new_chatify/domain-on-working/auth/repositories/auth_repository.dart';
 
-class SignUpUseCase extends UseCase<UserEntity, Params> {
+class SignInUseCase extends UseCase<UserEntity, Params> {
   final AuthRepository authRepository;
 
-  SignUpUseCase({required this.authRepository});
+  SignInUseCase({required this.authRepository});
 
   @override
   Future<Either<Failure, UserEntity>> call(Params params) async {
-    return await authRepository.signUp(params.email, params.password);
+    return await authRepository.signIn(params.email, params.password);
   }
 }
