+++ lib/domain-on-working/shared/entities/user_entitiy.dart
