+++ lib/injection_container.dart
@@ -1,11 +1,11 @@
 import 'package:get_it/get_it.dart';
-import 'package:new_chatify/features/auth/data/datasources/auth_datasource.dart';
-import 'package:new_chatify/features/auth/data/repositories/auth_repository_impl.dart';
-import 'package:new_chatify/features/auth/domain/repositories/auth_repository.dart';
-import 'package:new_chatify/features/auth/domain/usecases/signin_usecase.dart';
-import 'package:new_chatify/features/auth/domain/usecases/signout_usecase.dart';
-import 'package:new_chatify/features/auth/domain/usecases/signup_usecase.dart';
-import 'package:new_chatify/features/auth/presentation/bloc/bloc/auth_bloc.dart';
+import 'package:new_chatify/data-on-working/auth/datasources/auth_datasource.dart';
+import 'package:new_chatify/data-on-working/auth/repositories/auth_repository_impl.dart';
+import 'package:new_chatify/domain-on-working/auth/repositories/auth_repository.dart';
+import 'package:new_chatify/domain-on-working/auth/usecases/signin_usecase.dart';
+import 'package:new_chatify/domain-on-working/auth/usecases/signout_usecase.dart';
+import 'package:new_chatify/domain-on-working/auth/usecases/signup_usecase.dart';
+import 'package:new_chatify/presentation-on-working/auth/bloc/bloc/auth_bloc.dart';
 
 final sl = GetIt.instance;
