+++ lib/main.dart
@@ -3,9 +3,9 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:new_chatify/features/auth/presentation/bloc/bloc/auth_bloc.dart';
-import 'package:new_chatify/features/auth/presentation/pages/signin_view.dart';
 import 'package:new_chatify/injection_container.dart';
+import 'package:new_chatify/presentation-on-working/auth/bloc/bloc/auth_bloc.dart';
+import 'package:new_chatify/presentation-on-working/auth/pages/signin_view.dart';
 import 'package:new_chatify/presentation/bloc/chat/bloc/chat_bloc.dart';
 import 'package:new_chatify/presentation/bloc/user/user_bloc.dart';
 import 'package:new_chatify/firebase_options.dart';
