+++ lib/presentation-on-working/auth/bloc/bloc/auth_state.dart
