+++ lib/presentation-on-working/auth/widgets/re_logo_widget.dart
